+++ examples/typescript/simple_sponsored_transaction.ts
@@ -56,7 +56,7 @@ const example = async () => {
     feePayerAddress: sponsorAddress,
     data: {
       function: "0x1::aptos_account::transfer",
-      arguments: [bob.accountAddress, new U64(TRANSFER_AMOUNT)],
+      functionArguments: [bob.accountAddress, new U64(TRANSFER_AMOUNT)],
     },
   });