+++ examples/typescript/simple_transfer.ts
@@ -71,7 +71,7 @@ const example = async () => {
     data: {
       function: "0x1::coin::transfer",
       typeArguments: [new TypeTagStruct(StructTag.fromString(APTOS_COIN))],
-      arguments: [AccountAddress.fromHexInput(bob.accountAddress.toString()), new U64(TRANSFER_AMOUNT)],
+      functionArguments: [AccountAddress.fromHexInput(bob.accountAddress.toString()), new U64(TRANSFER_AMOUNT)],
     },
   });