+++ tests/e2e/api/transaction.test.ts
@@ -24,7 +24,7 @@ describe("transaction api", () => {
       sender: senderAccount.accountAddress.toString(),
       data: {
         function: "0x1::aptos_account::transfer",
-        arguments: [bob.accountAddress, new U64(10)],
+        functionArguments: [bob.accountAddress, new U64(10)],
       },
     });
     const authenticator = aptos.signTransaction({
@@ -49,7 +49,7 @@ describe("transaction api", () => {
         sender: senderAccount.accountAddress.toString(),
         data: {
           function: "0x1::aptos_account::transfer",
-          arguments: [bob.accountAddress, new U64(10)],
+          functionArguments: [bob.accountAddress, new U64(10)],
         },
       });
       const authenticator = aptos.signTransaction({